+++ Client/Config.cs
@@ -1,0 +1,13 @@
+﻿#if DEBUG
+#endif
+
+namespace SwitchPresence_Rewritten
+{
+    public class Config
+    {
+        public string IP, Client, BigKey, BigText, SmallKey, State;
+        public bool DisplayTimer, AllowTray;
+
+        public Config() { }
+    }
+}
